add: addmissing day UI issue, fix: WFH date issue
