+++ src/features/request/container/request.tsx
@@ -21,7 +21,6 @@ const Request: FC<IProjectsProps> = ({ uId }) => {
 	const [isVacational, setIsVacational] = useState(false);
 	const [isDisableDate, setIsDisableDate] = useState([]);
 	const [isUpcomingLeave, setIsUpcomingLeave] = useState<IUpComingLeave[]>();
-	const [isSuccess, setIsSuccess] = useState(false);
 
 	const REQUEST_ARR = [
 		{
@@ -179,8 +178,6 @@ const Request: FC<IProjectsProps> = ({ uId }) => {
 						setAddMissingDay({} as IAddMissingDay);
 					}}
 					isOpen={!isEmpty(addMissingDay)}
-					isSuccess={isSuccess}
-					setIsSuccess={setIsSuccess}
 				/>
 			)}
